Rename EpigenomesSorter component to EpigenomesLabels
